requirements.txt
requests==2.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
ShopifyAPI==12.4.0
APScheduler==3.10.4
loguru>=0.7.0
openpyxl==3.1.2 